requirements.txt
xhtml2pdf==0.2.5
djangorestframework==3.12.4
